refactor(api): merge dev related types
